requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
amqp==5.1.1
asgiref==3.6.0
async-timeout==4.0.2
asyncio==3.4.3
attrs==23.1.0
billiard==3.6.4.0
celery==5.2.7
certifi==2023.5.7
channels==4.0.0
charset-normalizer==3.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
coreapi==2.3.3
coreschema==0.0.4
Django==4.2.1
django-environ==0.10.0
django-filter==23.2
djangorestframework==3.14.0
drf-yasg==1.21.5
frozenlist==1.3.3
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
kombu==5.2.4
MarkupSafe==2.1.2
multidict==6.0.4
packaging==23.1
prompt-toolkit==3.0.38
psycopg==3.1.9
pytz==2023.3
redis==4.5.4
requests==2.30.0
ruamel.yaml==0.17.24
ruamel.yaml.clib==0.2.7
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.5.0
uritemplate==4.1.1
urllib3==2.0.2
vine==5.0.0
wcwidth==0.2.6
yarl==1.9.2
